:twisted_rightwards_arrows: Refactor of ManageOtc, ManageRx, OtcPage, and MedicinePage
