+++ src/Pages/OtcPage.tsx
@@ -20,6 +20,7 @@ import logButtonColor from "../utility/logButtonColor";
 import MedHistoryProvider from "../providers/MedHistoryProvider";
 import MedicineProvider from "../providers/MedicineProvider";
 import {updateDrugLog} from "./Common/updateDrugLog";
+import {updateMedicine} from "./Common/updateMedicine";
 
 interface IProps {
     activeTabKey: string | null,
@@ -142,22 +143,8 @@ const OtcPage = (props: IProps) => {
      */
     const handleMedicineEditModalClose = (drugInfo?: MedicineRecord | null) => {
         if (drugInfo && residentId) {
-            const drugData = {...drugInfo};
-
-            if (!drugData.Id) {
-                drugData.Id = null;
-            }
-
-            if (drugData.Notes === '') {
-                drugData.Notes = null;
-            }
-
-            if (drugInfo.Directions === '') {
-                drugData.Directions = null;
-            }
-
-            medicineProvider.post(drugData)
-            .then((drugRecord: any) => {
+            updateMedicine(medicineProvider, drugInfo)
+            .then((drugRecord) => {
                 RefreshOtcList(medicineProvider)
                 .then((drugList) => {
                     setOtcList(drugList).then(() => {});
